move health interval property to within healthconfig
